+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added `path.skew(x, y=0, center=(0, 0))`
 - Added `path.transform(transform, center=(0, 0))`
 - Added `path.arc(center, radius, startAngle, endAngle, clockwise)`
+- Added `path.arcTo(point1, point2, radius)`
 
 ## [0.2.0] - 2020-09-26